I tried the optimization to sort by floats, which gave a higher maximum, but it's still not getting to the correct answer yet. Another idea would be to sort the ties (if there will be any) by increasing weight.

+++ Python/Stanford-Algo/Part-2/Prob-03-knapsack/optimized_knapsack_02.py
@@ -36,7 +36,7 @@ def ReadValuesAndRatios( fileName ):
             edge = line.split()
             value = int(edge[0])
             weight = int(edge[1])
-            value_ratio = value/weight
+            value_ratio = float( float(value)/float(weight) )
             elements.append( (value_ratio, weight, value) )
 
     # TOTO sort by descending the ratio fist, but then ascending then weight.
